+++ src/components/modelRender.go
@@ -186,7 +186,7 @@ func MetaDataRender(m model) string {
 	// process bar
 	metaDataBar := ""
 	if len(m.fileMetaData.metaData) == 0 && len(m.fileModel.filePanels[m.filePanelFocusIndex].element) > 0 && !m.fileModel.renaming {
-		m = returnMetaData(m)
+		m = ReturnMetaData(m)
 	}
 
 	maxKeyLength := 0